Fix regression for resources with property (step with :<)
Jena Model.listStatements is not working as documented with a NULL
argument in place of the subject. This applies for in-memory models as
well as TDB models; Java 8 as well as Java 11/12. Will consider
submitting bug report for this. In the meantime
Model.listResourcesWithProperty is a more appropriate method to use, and
works fine. Using Model.listObjectOfProperty for outgoing edges as well.