Getting graph name from document rather than json
In general, all documents have a single root object that holds the
IRI used for the named graph that gets imported into Jena. Documents
are expected to have exactly one of these. Rather than assume that
every message is in JSON, assume instead that the parsed model
contains the expected object type, and use the IRI from that as the
document name.