Change in write txn, updates to perform model validation;
Write transactions are now handled differently. previously,
transaction commit and abort commands were written into the write-tx
macro. The method to roll back a commit that needed to be aborted
was to raise an exception, which is too heavy handed and un-clojure
like.

Now code that calls the macro is explicity expected to commit
the transaction. This means that an exception will be raised if the
caller fails to commit or abort. At the moment, the only way data gets
into the database is via the load-model function in database.load, so
this is a relatively easy change to make.

Additionally model validation is moving into the core database handling, rather
than at the edge in the sink namespace. SHACL validaiton feels more
like a core feature of the datastore.