Adding 'type' param to query, currently only supports ask
This is necessary to specify the type of query to execute when a query
is defined by algebra alone; otherwise it seems to default to select,
and query type is not inferred from the algebra itself, at least not
for ask. This makes it possible to construct an ask query using just algebra.