Changing the way the resolver cache expires resources
Previously only cache entries tied to specific subjects
identified for inclusion in a model were expired, such as
genes and diseases. This left out cases where entities
included in the model itself are updated (such as the disease
for a gene validity curation). This code harvests all the
referenced resources in a new model and expires any references
to all of them in the resolver cache, addressing the issue.