Added allowances for assertion_ prefix in GV iris
Per Phil's request on Slack:

Phillip Weller
  16 days ago
I'm probably imagining things - but didn't genegraaph used to return the current GVD record if you passed in an identifier without the timestamp?

8 replies

Tristan Nelson
  16 days ago
It does, but it doesn't seem to be happy if you add the assertion_ prefix. This link works, for example: https://search.clinicalgenome.org/kb/gene-validity/CGGV:38729563-bf36-48ae-929e-fa69a225de39

Phillip Weller
  16 days ago
How difficult would it be to also honor the current string format?

Tristan Nelson
  16 days ago
Including the assertion_ prefix before the UUID, you mean?

Phillip Weller
  16 days ago
Yes, the one currently in use.

Tristan Nelson
  16 days ago
I don't think it would be too bad. Is there something specific you're trying to accomplish with a particular timeframe?

Phillip Weller
  16 days ago
It would be nice if it could do maybe by the 15th of march?  Thats when we'll be freezing the code for the Q1 update.  Trying to implement a request for Jonathan and it requires always accessing the latest of whatever assertion he bookmarked.

Tristan Nelson
  16 days ago
That seems feasible, will let you know if that turns out to be a problem.

Phillip Weller
  16 days ago
Thanks!